+++ packages/form/src/Textarea/Textarea.js
@@ -60,7 +60,7 @@ const Textarea = (props: Props) => {
   return (
     <div className={textareaClass} style={style} data-qa={qa}>
       {label && <InputLabel htmlFor={id}>{label}</InputLabel>}
-      <textarea id={id} placeholder={placeholder} name={name} value={value} onChange={onChange} onBlur={onBlur} disabled={disabled}></textarea>
+      <textarea id={id} placeholder={placeholder} name={name} value={value} onChange={onChange} onBlur={onBlur} disabled={disabled} required={required}></textarea>
     </div>
   )
 }