+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 ### Added
 - Added select disabled state
+- `Select` add type
 
 ### Fixed
+- `Textarea` Fixed an issue where required property was not being passed properly
 - `AccordeonTab` Added an option to pass a custom toggle function so that state can be managed in parent components.